feat: 아이콘 정리
