+++ src/components/icons/BtnIcons.tsx
@@ -14,3 +14,27 @@ export const SettingIcon = () => {
     </svg>
   );
 };
+
+export const GoBackIcon = () => {
+  return (
+    <>
+      <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path
+          d="M14.5 5L7 12.5L14.5 20"
+          stroke="#242424"
+          strokeWidth="1.5"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        />
+      </svg>
+    </>
+  );
+};
+
+export const DelIcon = () => {
+  return (
+    <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M2 2L10 10M2 10L10 2" stroke="#B3B3B3" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
+    </svg>
+  );
+};
