--- src/components/icons/DelIcon.tsx
@@ -1,9 +1,0 @@
-const DelIcon = () => {
-  return (
-    <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path d="M2 2L10 10M2 10L10 2" stroke="#B3B3B3" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-};
-
-export default DelIcon;
