--- src/components/icons/GoBackIcon.tsx
@@ -1,15 +1,0 @@
-export const GoBackIcon = () => {
-  return (
-    <>
-      <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <path
-          d="M14.5 5L7 12.5L14.5 20"
-          stroke="#242424"
-          strokeWidth="1.5"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    </>
-  );
-};
