+++ src/components/navigations/goBack.tsx
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled';
 import { useRouter } from 'next/router';
 
-import PATH from '../../constants/path';
-import { GoBackIcon } from '../icons/GoBackIcon';
+import { GoBackIcon } from '../icons/BtnIcons';
 import TopNavBar from './navigation_top';
 
 export default function GoBack() {
